Added files via upload
created a condition for each of the instructions that stores each element of an instruction in a corresponding string. 
stores the labels in a symbol table

the conditions hardcode the opcode and funct

need to go through each condition and convert each register to its decimal value then to its binary value as a string (a link is provided to a function that converts unsigned ints to binary strings).  http://www.cplusplus.com/forum/general/65862/
also need to convert the immediate values from strings to unsigned ints then to binary string as well.

beq, bne, j, and jal will be tricky as they have labels in them

sw and lw will also be difficult to convert because of the offsets